requirements.txt
playwright==1.41.2
pandas==2.2.0
openpyxl==3.1.2
PyYAML==6.0.1
